Fixed JFile and JFolder delete functions for restrictive persmissions
git-svn-id: http://joomlacode.org/svn/joomla/development/trunk@5245 6f6e1ebd-4c2b-0410-823f-f34bde69bce9

+++ libraries/joomla/filesystem/folder.php
@@ -191,13 +191,18 @@ class JFolder
 			$ftp = & JFTP::getInstance($config->getValue('config.ftp_host'), $config->getValue('config.ftp_port'));
 			$ftp->login($config->getValue('config.ftp_user'), $config->getValue('config.ftp_pass'));
 
+	    }
+	    // In case of restricted permissions we zap it one way or the other
+	    // as long as the owner is either the webserver or the ftp
+	    if(@rmdir($path)){
+	    	$ret = true;
+	    } else {
 			// Translate path and delete
 			$path = JPath::clean(str_replace(JPATH_SITE, $ftpRoot, $path));
 			$ret = $ftp->delete($path);
+	    }
+	    if($ftpFlag){
 			$ftp->quit();
-		} else {
-			// Do it the regular way
-			$ret = rmdir($path);
 		}
 		return $ret;
 	}
